done with login/signup styling, corrected link in axiosWithAuth, added logo in both pages,corrected localstorage id value
