+++ src/components/Login/Login.css
@@ -5,28 +5,34 @@
     max-width: 580px;
     max-height: 490px;
     margin: 0 auto;
-    margin-top: 20%;
+    margin-top: 5%;
     padding: 75px;
     border: 1px solid rgba(0, 0, 0, 0.12);;
 }
 
+.ptlogo {
+    width: 60%;
+    margin: 0 auto 8% auto;
+    filter: invert(52%) sepia(79%) saturate(2930%) hue-rotate(22deg) brightness(97%) contrast(91%);
+}
+
 h1.loginHeader {
     text-align: center;
-    margin-bottom: 8%;
-    color: #4f23ef;
+    margin-bottom: 6%;
+    color: #B8860C;
 }
 
 .loginInput {
     margin-bottom: 3%;
-    border-color:#4f23ef ;
-    font-size: 1.6rem;
+    border-color:#B8860C;
+    font-size: 1.2rem;
 }
 
 .loginButton {
-    background-color:#4f23ef;
+    background-color:#B8860C;
     color: white;
     border: none;
-    min-width: 120px;
+    min-width: 130px;
     font-family: 'Roboto', sans-serif;
-    font-size: 1.2rem;
+    font-size: 1.3rem;
 }
