--- src/components/SignUp.js
@@ -1,108 +1,0 @@
-import React, { useState, useContext } from 'react';
-import { Form, FormGroup, Label, Input,Button } from 'reactstrap';
-import axios from "axios";
-import UserContext from '../contexts/UserContext';
-
-const initialState = {
-    username: "",
-    password: "",
-	  // email: "",
-    firstname:"",
-    lastname:""
-};
-
-const SignUp = (props) => {
-    const [newUser, setNewUser] = useState(initialState);
-    const user = useContext(UserContext);
-
-    const changeHandler = (event) => {
-        setNewUser({
-            ...newUser,
-            [event.target.name]: event.target.value
-        });
-    };
-
-    const submitForm = (event) => {
-        event.preventDefault();
-        axios.post("https://tt-webpt-92-potluck-app.herokuapp.com/api/register", newUser)
-          .then((res) => {
-            user.username = newUser.username;
-            localStorage.setItem("username", newUser.username);
-            console.log(res.data);
-            props.history.push("/Home");
-          })
-          .catch((err) => {
-            console.error("something went wrong: ", err);
-          })
-    };
-
-    return (
-        <>
-          <Form onSubmit={submitForm}>
-            <FormGroup>
-              <Label for="username">Username</Label>
-              <Input
-                type="text"
-                name="username"
-                id="username"
-                placeholder="enter your username"
-                value={newUser.username}
-                onChange={changeHandler}
-                required
-              />
-            </FormGroup>
-            <FormGroup>
-              <Label for="password">Password</Label>
-              <Input
-                type="password"
-                name="password"
-                id="password"
-                placeholder="enter your password"
-                value={newUser.password}
-                onChange={changeHandler}
-                required
-              />
-            </FormGroup>
-            <FormGroup>
-                    <Label for="firstname">First name</Label>
-                    <Input
-                      type="text"
-                      name="firstname"
-                      id="firstname"
-                      placeholder="enter your first name"
-                      value={newUser.firstname}
-                      onChange={changeHandler}
-                      required
-                    />
-            </FormGroup>
-            <FormGroup>
-                    <Label for="lastname">Last name</Label>
-                    <Input
-                      type="text"
-                      name="lastname"
-                      id="lastname"
-                      placeholder="enter your last name"
-                      value={newUser.lastname}
-                      onChange={changeHandler}
-                      required
-                    />
-            </FormGroup>
-            {/* <FormGroup>
-              <Label for="email">Email</Label>
-              <Input
-                type="email"
-                name="email"
-                id="email"
-                placeholder="enter your email"
-                value={newUser.email}
-                onChange={changeHandler}
-                required
-              />
-            </FormGroup> */}
-            <Button type="submit">Submit</Button>
-          </Form>
-        </>
-      );
-};
-
-export default SignUp;
