+++ src/utils/axiosWithAuth.js
@@ -4,7 +4,7 @@ export const axiosWithAuth = () => {
   const token = localStorage.getItem('token')
 
   return axios.create({
-    baseURL: "https://potluckplanner2021.vercel.app",
+    baseURL: "https://tt-webpt-92-potluck-app.herokuapp.com",
     headers: {
       authorization: token
     }
